Add type to global variable

+++ src/Tether_03.jl
@@ -2,9 +2,9 @@
 # initially moving upwards with 4 m/s.
 using ModelingToolkit, OrdinaryDiffEq, PyPlot, LinearAlgebra
 
-G_EARTH  = Float64[0.0, 0.0, -9.81]    # gravitational acceleration [m/s²]
-L0 = -10.0                             # initial spring length      [m]
-V0 = 4                                 # initial velocity           [m/s]
+G_EARTH     = Float64[0.0, 0.0, -9.81]          # gravitational acceleration [m/s²]
+L0::Float64 = -10.0                             # initial spring length      [m]
+V0::Float64 = 4                                 # initial velocity           [m/s]
 
 # model, Z component upwards
 @parameters mass=1.0 c_spring0=50.0 damping=0.5 l0=L0
